Refactor so all EnhancedEditor events are in a single write action
Also, use the NON_MODAL modality to invoke them later, rather than
using the UIThreadScheduler. This is sorta hacks based on looking
at the source, but it's not unreasonable.

When already on the UI thread, invokeAndWait drops the modality
entirely, so TransactionGuard doesn't wrap it to mark
`myWritingAllowed=true`. We use NON_MODAL modality instead of
making our own ModalityStateEx containing the buffer or something,
for example, because the modality state we provide is not added to
the `myWriteSafeModalities` map for some reason. We could figure
out how to do that by hand, but NON_MODAL is always a member, so....

The other advantage of doing all events in a single write action is
that there's less possible UI jank.