Refactor Notification *again* to better support events
Yes, the underlying data format must be a List, but unless you
actually want to handle arguments as a list of Any, this change
will let us provide an array-formatted class whose fields will
be deserialized for us. If you do want a homogeneous list of
args, you can use the new NotificationListPacket type. We use
it for the builtin RedrawEvent currently.

There's also some awfully gross reflection magic to make the
BufferEventArg handling work as expected. It'd be great to
refactor all that stuff, but it works well enough, and I've added
a few regression tests as well to make sure it stays sane.