Add a (somewhat hamfisted) way to interact with completion popup
Eventually we'll want to refactor this so that users can customize
the interaction via vimrc mappings, but this will work for proof
of concept for now. More immediately pressing is to take the changes
to the document that IntelliJ makes and pass them back to nvim.

An alternate approach would be to hook into *vim's* text completion
APIs and pull results from IntelliJ, but I don't think that will be
significantly less hacky, and we will need to update nvim from
IntelliJ eventually anyway, in order to support refactorings, etc.